feat: encode api and collab token with props
